plot_rt_app: Fix frequency plot
We use df_util to define the xlimit of the frequency plot to ensure we
have an aligned view. But after last code reshuffle we process df_utils
after frequency and end up plotting in the wrong xlimits.

Fix that by reordering the processing such that df_util is done first.

We still plot frequency above util in the final result.png image.

Signed-off-by: Qais Yousef <qyousef@layalina.io>