start_perfetto: Allow different perfetto configs
That each test can optionally specify. If not specified we use the
default.

The default can be too verbose for long tests, so we need something
custom to better manage the size.

Later we will give the user the option to select the verbosity when
starting the test. In case they want to capture a single run with all
the details to help investigate something.

Increase the default config.pbtx size to 2GiB since we will allow custom
configs to capture smaller data for some tests. 1GiB could easily be
filled for a single PCMark run for example.

Signed-off-by: Qais Yousef <qyousef@layalina.io>