plot_frequency: Only print clusters
The code was heavily refactored too. And the query was fixed as it
didn't work against newest captures from Android devices. We need to be
specific which column we want from the database.

With new changes, we don't unnecessarily plot other cpus in the same
cluster as they'll be identical.

We still have issues in specifying the correct figure size in case there
were too many clusters or a large number of cpus in case they aren't
clustered.

Also fix SettingWithCopyWarning.

Signed-off-by: Qais Yousef <qyousef@layalina.io>