plot/freq: Improve plot() signature
Use row_pos instead of start_row, as it's a better name.

Take num_rows as an argument too, otherwise using the wrong number
of rows in subplot() produces bad results. We must match what the caller
uses, hence the need to pass it.

And take a list of cpus to plot, to allow caller to limit what cpus we
plot.

Signed-off-by: Qais Yousef <qyousef@layalina.io>