plot_sysbench: Use line plot again
The bar was too dense for high n umber of iterations. It's easier to see
the trends with a line plot.

Also remove the text annotation from the mean axhline. The calculation
of the position wasn't right. We can see the actual values from the
table, so remove them for now until we fix the text.plot() to better
manage adding it at the right location.

Signed-off-by: Qais Yousef <qyousef@layalina.io>