plot/wa_results: Make width depend on num_rows too
Square looks better when zoomed in. There seem to be room for
improvement still as the file is large. Maybe my viewers were bad and we
could do with smaller size. A multiplier of 2 instead of 4 produced
results that tight_layout() didn't complain about being too crammed, but
it didn't feel much readable, in my viewer at least.

Signed-off-by: Qais Yousef <qyousef@layalina.io>