plot/wa_results: Never plot in a single row
The condition was wrong, we should have used len(metrics), NOT
num_metrics which is the number of rows we calculated!

We could try to get num_cols automatically handled for when len(metrics)
is less than 4, but it's not worth it. Just always have this static
division until the need to do otherwise arises :-)

Signed-off-by: Qais Yousef <qyousef@layalina.io>