Workaround blueocean bug when running parallel stages
When dex2oat stage is not run (it's optional), the UI gets stuck showing
'waiting for run to start'. This seems to be a bug in blueocean, see
this post

	https://issues.jenkins.io/browse/JENKINS-58085

As a workaround suggested in that thread, adding a dummy stage that
finishes quickly restores the usual UI behavior.

This was observed in blueocean version 1.25.3. The problem has been
there for a while..

Signed-off-by: Qais Yousef <qyousef@layalina.io>