run_dex2oat: Finally use the env.ITERATIONS and env.DELAY
Follow what the user asks for.. :-)

Also improve sched-tests-dex2oat to include the post { always {} } step
to ensure we clean up in case something failed. It's hard to kill the
command to force JIT compilation, at least let us wait to ensure we
leave the device in a good known state.

Fix the description in the config.xml since now we will NOT trigger
dex2oat only once, we'll follow the number of iters and delay specified
by the user.

Signed-off-by: Qais Yousef <qyousef@layalina.io>