Set parameters from jenkinsfile for android tests
This will make it easier to add and remove params without having to
constantly update the config.xml files. Though the job must run once
everytime a parameter is added/removed for jenkins to pick up the new
change. We try to catch that and throw an error if a parameter is
missing.

While at it, add a parameter to set job description instead of having to
do this manually after the run.

We'll convert others later too.

Signed-off-by: Qais Yousef <qyousef@layalina.io>