Remove parameters from all jenkinsfiles except for NODE
Now that we define parameters in jenkinsfile, we can keep the config.xml
simple and easier to replicate.

We must keep the NODE so that the first run the user can still specify
a correct node, but it'll likely to fail due to parameters need a single
dummy run for jenkins to pickup the parameters from jenkinsfile.

Signed-off-by: Qais Yousef <qyousef@layalina.io>