vars/verify_params: Allow DESCRIPTION to be empty
The test doesn't actually check whether a variable exists, it checks if
it is empty too. Since we can't differentiate between the two until we
find a better way to do the check, omit checking for DESCRIPTION as it
being empty is fine but we were getting erroneous aborts when it was
empty.

Signed-off-by: Qais Yousef <qyousef@layalina.io>