plot/pelt: Filter out overutilized
We can get a sequence of short bursts of overutilized switching between
on and off. Which ends up creating noisy graphs as the span width will
overlap heavily. To alleviate that, we filter based on the total width
of the x-axis. It should be greater than 1/1000th of that for us to
attempt to mark an area as overutilized.

Signed-off-by: Qais Yousef <qyousef@layalina.io>