start_perfetto: Collect uclamp when COLLECT_PELT is true
The 2 information are interrelated. Until we have a good reason why we
should collect them each on isolation, use the same COLLECT_PELT flag to
collect them together optionally. This flag is set because the user
would like more info to debug something, it makes sense to collect both
data together.

Size could be one potential issue to split the 2. Current use case
requires them both to be collected though. It is expected low number of
iterations to be selected when COLLECT_PELT is set as it bloats the
trace considerably.

Signed-off-by: Qais Yousef <qyousef@layalina.io>