start_perfetto: Fix COLLECT_{PELT,UCLAMP} is not always defined
Don't use explicit binding and try to access it as env variable instead
so that it's empty if not defined. The ${} form assumes explicit binding
by jenkins pipeline and if a job doesn't define it, which not all jobs
do, we'll get a weird exception about undefined binding for this
variable.

Signed-off-by: Qais Yousef <qyousef@layalina.io>