run_wa.groovy: Force run the script even if wa_output exists
If a build was aborted, workspace might not have been deleted.

Deleting workspace before we start is a problem as we end up deleting
the agendas etc that we carry as part of the test which gets checked out
automatically whenever we run.

For now, pass '-f' flag to force the run. Would be better to ensure the
workspace is properly cleaned though.

Signed-off-by: Qais Yousef <qyousef@layalina.io>