sched-tests-profile-android: Revive this test
The goal is to run multiple android tests and collect all the data in
one go rather than manually trigger them all. We should end up with
a report.

We setup the infrastructure in this commit.

We have a new deps on Copy Artifact Plugin. Parent 'myci' project
should provide this for us.

We don't archive perfetto-traces or tar files, only pictures. We should
improve on this to reach the ultimate goal of having a final report
produce to examine.

Signed-off-by: Qais Yousef <qyousef@layalina.io>