plotting: Add support to save processed data into csv files
Captured data from perfetto could be useful to keep around for more
post-processing later. And to archive.

This change provides a new save_csv(prefix) function for each
post-processing module that acts on perfetto-trace directly.

We refactor the code along side it to reduce repeated code blocks that
tries to populate the dataframe once.

We don't populate them at init because we want to avoid getting the data
frame while the trace is opened. When it is opened the RAM is already
stressed. Obtaining the dataframes after closing it improves performance
and reduces memory pressure. Callers close the trace file after calling
init() functions on all modules.

Signed-off-by: Qais Yousef <qyousef@layalina.io>