vars/plot*: Prevent multiple plotting functions to run in parallel
Processing perfetto traces can be taxing on the system. If we run
multiple jobs in parallel we could end up killing the system.

Use flock to serialize plotting function. Will slow things down, but
ensure we don't kill the system, which is worse.

Generally if a single job is running, this should have no impact as we
serialize multiple plots in this case anyway.

Signed-off-by: Qais Yousef <qyousef@layalina.io>