perfetto, plotting: Handle breakage of trace_processor_shell
Newer versions of trace_processor_shell which perfetto python interface
automatically pulls requires specific extensions on x86: SSE4.2, BMI2
and AVX2. But on VirgualBox VM, it is not possible to get BMI2 which
causes processing the perfetto trace impossible in these environments.

We can workaround it by using a local install of v25 which doesn't have
this problem.

Update the code so that we use a local install of trace_processor_shell
if found. Otherwise revert to using the prebuilts that perfetto
downloads automatically.

Signed-off-by: Qais Yousef <qyousef@layalina.io>